refactor(ui): split useKeyboard.ts into 7 focused modules (1121→84 lines)
Extract types, state helpers, and 5 sub-hooks from useKeyboard.ts.

Modules extracted:
- keyboard-types.ts: KeyboardState, BulkKeyEntry, emptyState(), SetState, KeyboardRefs
- keyboard-state-helpers.ts: parseDefinitionLayout() shared helper
- useKeyboardReload.ts: 9-phase device HID reload workflow
- useKeyboardLoaders.ts: loadDummy(), loadPipetteFile()
- useKeyboardSetters.ts: keymap, encoder, macro, dynamic entry setters
- useKeyboardLighting.ts: backlight, RGBlight, VialRGB, QMK settings
- useKeyboardPersistence.ts: serialize, apply, reset, pipette-file QMK

Orchestrator owns all state/refs and passes to sub-hooks.
BulkKeyEntry/KeyboardState re-exported for backward compatibility.
